패키지디렉토리 도메인형으로 변경
