+++ src/main/java/com/multi/quizwiki/interceptor/PboardInterceptor.java
@@ -6,7 +6,7 @@ import javax.servlet.http.HttpServletResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.servlet.HandlerInterceptor;
 
-import com.multi.quizwiki.service.PboardService;
+import com.multi.quizwiki.pboard.service.PboardService;
 
 import lombok.NoArgsConstructor;
