+++ src/main/java/com/multi/quizwiki/pboard/dao/ProblemDAO.java
@@ -1,14 +1,14 @@
-package com.multi.quizwiki.dao;
+package com.multi.quizwiki.pboard.dao;
 
 import java.io.IOException;
 import java.util.List;
 
 import org.springframework.web.multipart.MultipartFile;
 
-import com.multi.quizwiki.entity.problem.PrintFileEntity;
-import com.multi.quizwiki.entity.problem.ProblemCateEntity;
-import com.multi.quizwiki.entity.problem.ProblemChoiseEntity;
-import com.multi.quizwiki.entity.problem.ProblemEntity;
+import com.multi.quizwiki.pboard.entity.PrintFileEntity;
+import com.multi.quizwiki.pboard.entity.ProblemCateEntity;
+import com.multi.quizwiki.pboard.entity.ProblemChoiseEntity;
+import com.multi.quizwiki.pboard.entity.ProblemEntity;
 
 public interface ProblemDAO {
 	public ProblemEntity insert(ProblemEntity problem);
