+++ src/main/java/com/multi/quizwiki/pboard/entity/PrintFileEntity.java
@@ -1,4 +1,4 @@
-package com.multi.quizwiki.entity.problem;
+package com.multi.quizwiki.pboard.entity;
 
 import java.sql.Timestamp;
 
@@ -8,8 +8,6 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.Table;
 
-import org.hibernate.annotations.UpdateTimestamp;
-
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -18,24 +16,21 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 @Entity
-@Table(name = "solv")
-public class SolvEntity {
+@Table(name = "print_file")
+public class PrintFileEntity {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	String solvId;
-	int problemId;
-	String memberId;
-	String solvAnswer;
-	@UpdateTimestamp 
-	Timestamp solvEditDate;
+	String printFileId;
+	String problemId;
+	String printfileOrigin;
+	String printfileStore;
 	
 	
-	public SolvEntity(int problemId, String memberId, String solvAnswer) {
+	public PrintFileEntity(String problemId, String printfileOrigin, String printfileStore) {
 		this.problemId = problemId;
-		this.memberId = memberId;
-		this.solvAnswer = solvAnswer;
+		this.printfileOrigin = printfileOrigin;
+		this.printfileStore = printfileStore;
 	}
-	
 	
 	
 }
