+++ src/main/java/com/multi/quizwiki/pboard/repository/PboardCateRepository.java
@@ -1,8 +1,8 @@
-package com.multi.quizwiki.repository;
+package com.multi.quizwiki.pboard.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.multi.quizwiki.entity.pboard.PboardCateEntity;
+import com.multi.quizwiki.pboard.entity.PboardCateEntity;
 
 public interface PboardCateRepository extends JpaRepository<PboardCateEntity, String>{
