+++ src/main/java/com/multi/quizwiki/pboard/repository/PboardReplyRepository.java
@@ -1,10 +1,10 @@
-package com.multi.quizwiki.repository;
+package com.multi.quizwiki.pboard.repository;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.multi.quizwiki.entity.pboard.PboardReplyEntity;
+import com.multi.quizwiki.pboard.entity.PboardReplyEntity;
 
 public interface PboardReplyRepository extends JpaRepository<PboardReplyEntity, String>{
 	public Page<PboardReplyEntity> findByPboardId(String pboardId, Pageable pageable);
