+++ src/main/java/com/multi/quizwiki/pboard/repository/PboardRepository.java
@@ -1,11 +1,11 @@
-package com.multi.quizwiki.repository;
+package com.multi.quizwiki.pboard.repository;
 
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.multi.quizwiki.entity.pboard.PboardEntity;
+import com.multi.quizwiki.pboard.entity.PboardEntity;
 
 public interface PboardRepository extends JpaRepository<PboardEntity, String>{
 	public Page<PboardEntity> findByPboardCateIdAndPboardStatusNot(int pboardCateId,String pboardStatus, Pageable pageable);
