+++ src/main/java/com/multi/quizwiki/pboard/repository/PrintFileRepository.java
@@ -1,8 +1,8 @@
-package com.multi.quizwiki.repository;
+package com.multi.quizwiki.pboard.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.multi.quizwiki.entity.problem.PrintFileEntity;
+import com.multi.quizwiki.pboard.entity.PrintFileEntity;
 
 public interface PrintFileRepository extends JpaRepository<PrintFileEntity, String>{
 	public PrintFileEntity findByProblemId(String problemId);
