+++ src/main/java/com/multi/quizwiki/pboard/repository/ProblemCateRepository.java
@@ -1,8 +1,8 @@
-package com.multi.quizwiki.repository;
+package com.multi.quizwiki.pboard.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.multi.quizwiki.entity.problem.ProblemCateEntity;
+import com.multi.quizwiki.pboard.entity.ProblemCateEntity;
 
 public interface ProblemCateRepository extends JpaRepository<ProblemCateEntity, String>{
