+++ src/main/java/com/multi/quizwiki/pboard/repository/ProblemChoiseRepository.java
@@ -1,8 +1,8 @@
-package com.multi.quizwiki.repository;
+package com.multi.quizwiki.pboard.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.multi.quizwiki.entity.problem.ProblemChoiseEntity;
+import com.multi.quizwiki.pboard.entity.ProblemChoiseEntity;
 
 public interface ProblemChoiseRepository extends JpaRepository<ProblemChoiseEntity, String>{
 	public void deleteAllByProblemId(String problemId);
