+++ src/main/java/com/multi/quizwiki/pboard/repository/ProblemRepository.java
@@ -1,10 +1,10 @@
-package com.multi.quizwiki.repository;
+package com.multi.quizwiki.pboard.repository;
 
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.multi.quizwiki.entity.problem.ProblemEntity;
+import com.multi.quizwiki.pboard.entity.ProblemEntity;
 
 public interface ProblemRepository extends JpaRepository<ProblemEntity, String>{
 	public List<ProblemEntity> findAllByPboardId(String pboardId);
