+++ src/main/java/com/multi/quizwiki/pboard/service/PboardServiceImpl.java
@@ -1,4 +1,4 @@
-package com.multi.quizwiki.service;
+package com.multi.quizwiki.pboard.service;
 
 import java.io.IOException;
 import java.util.List;
@@ -11,15 +11,15 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import com.multi.quizwiki.common.FileUploadLogicService;
-import com.multi.quizwiki.dao.PboardDAO;
-import com.multi.quizwiki.dao.ProblemDAO;
-import com.multi.quizwiki.entity.pboard.PboardCateEntity;
-import com.multi.quizwiki.entity.pboard.PboardEntity;
-import com.multi.quizwiki.entity.pboard.PboardReplyEntity;
-import com.multi.quizwiki.entity.problem.PrintFileEntity;
-import com.multi.quizwiki.entity.problem.ProblemCateEntity;
-import com.multi.quizwiki.entity.problem.ProblemChoiseEntity;
-import com.multi.quizwiki.entity.problem.ProblemEntity;
+import com.multi.quizwiki.pboard.dao.PboardDAO;
+import com.multi.quizwiki.pboard.dao.ProblemDAO;
+import com.multi.quizwiki.pboard.entity.PboardCateEntity;
+import com.multi.quizwiki.pboard.entity.PboardEntity;
+import com.multi.quizwiki.pboard.entity.PboardReplyEntity;
+import com.multi.quizwiki.pboard.entity.PrintFileEntity;
+import com.multi.quizwiki.pboard.entity.ProblemCateEntity;
+import com.multi.quizwiki.pboard.entity.ProblemChoiseEntity;
+import com.multi.quizwiki.pboard.entity.ProblemEntity;
 
 import lombok.NoArgsConstructor;
